Delete test/e2e and rename test/e2eclusterctl package to test/e2e
