+++ docs/development.md
@@ -350,7 +350,7 @@ You can optionally set the following variables:
 | `AZURE_SSH_PUBLIC_KEY_FILE`        | Use your own SSH key.                                                                                                     | `${GOPATH}/src/sigs.k8s.io/cluster-api-provider-azure/.sshkey`                                  |
 | `AZURE_CONTROL_PLANE_MACHINE_TYPE` | The machine type of the control plane.                                                                                    | `Standard_D2s_v3`                                                                               |
 | `AZURE_NODE_MACHINE_TYPE`          | The machine type of the node machine.                                                                                     | `Standard_D2s_v3`                                                                               |
-| `E2E_CONF_FILE`                    | The path of the [E2E configuration file](https://cluster-api.sigs.k8s.io/developer/e2e.html#defining-an-e2e-config-file).      | `${GOPATH}/src/sigs.k8s.io/cluster-api-provider-azure/test/e2eclusterctl/config/azure-dev.yaml`  |
+| `E2E_CONF_FILE`                    | The path of the [E2E configuration file](https://cluster-api.sigs.k8s.io/developer/e2e.html#defining-an-e2e-config-file).      | `${GOPATH}/src/sigs.k8s.io/cluster-api-provider-azure/test/e2e/config/azure-dev.yaml`            |
 | `E2E_ARTIFACTS`                    | The path of the artifacts folder.                                                                                         | `${GOPATH}/src/sigs.k8s.io/cluster-api-provider-azure/_artifacts`                               |
 | `E2E_SKIP_RESOURCE_CLEANUP`        | Set to `true` if you do not want the bootstrap and workload clusters to be cleaned up after running E2E tests.            | `false`                                                                                         |
 | `E2E_USE_EXISTING_CLUSTER`         | Set to `true` if you are using your own bootstrap cluster.                                                                | `false`                                                                                         |
