+++ test/e2e/data/infrastructure-azure/cluster-template.yaml
